Rename: note_player -> music_player
